IMR-173 refactor: move config, crawler, db, migraion dir in src and modi import path
