+++ src/config/__init__.py
@@ -1,0 +1,3 @@
+from .config import DBConfig, GenieConfig
+
+__all__ = ["DBConfig", "GenieConfig"]
