+++ src/crawler/album/__init__.py
@@ -1,0 +1,1 @@
+from .album_info_crawler import crawlAlbumInfo
