+++ src/crawler/album_main.py
@@ -1,12 +1,13 @@
 import pandas as pd
 from tqdm import tqdm
-from config.config import GenieConfig
+from ..config import GenieConfig
 from tqdm.contrib.logging import logging_redirect_tqdm
-from .src.utils import getLogger
-from .album.album_info_crawler import crawlAlbumInfo
+from .utils import getLogger
+from .album import crawlAlbumInfo
 
 ALBUM_DETAIL_URL = "https://www.genie.co.kr/detail/albumInfo?axnm="
 config = GenieConfig()
+
 
 def crawlAlbum(album_ids: list[str]) -> pd.DataFrame:
     # log file setting
@@ -17,7 +18,7 @@ def crawlAlbum(album_ids: list[str]) -> pd.DataFrame:
     log.info("Start album info crawling ...")
     album_infos = []
     failed_crawling_album_ids = []
-    
+
     for album_id in tqdm(album_ids, position=0, leave=True):
         with logging_redirect_tqdm():
             try:
@@ -33,4 +34,3 @@ def crawlAlbum(album_ids: list[str]) -> pd.DataFrame:
     album_infos_df = pd.DataFrame(album_infos)
 
     return album_infos_df
-
