+++ src/crawler/playlist/__init__.py
@@ -1,0 +1,1 @@
+from .playlist_crawler import getPlaylistInfo
