+++ src/crawler/song/song_detail/__init__.py
@@ -1,0 +1,4 @@
+from .song_daliy_chart_crawler import *
+from .song_info_crawler import *
+from .song_like_crawler import *
+from .song_lyrics_crawler import *
