+++ src/crawler/song/song_detail/song_info_crawler.py
@@ -1,4 +1,4 @@
-from ...src.utils import tranlateSongInfoAttrToEng
+from ...utils import tranlateSongInfoAttrToEng
 from bs4 import BeautifulSoup, Tag, ResultSet
 
 SONG_INFO_DATA_SELECTOR = "ul.info-data"
