+++ src/crawler/song/song_detail/song_like_crawler.py
@@ -1,5 +1,5 @@
 from bs4 import BeautifulSoup
-from ...src.utils import txt2int
+from ...utils import txt2int
 
 SONG_LIKE_SELECTOR = "a.like em#emLikeCount"
