+++ src/crawler/song_main.py
@@ -1,12 +1,13 @@
 import pandas as pd
 from tqdm import tqdm
 from tqdm.contrib.logging import logging_redirect_tqdm
-from config.config import GenieConfig
-from .src.utils import getLogger
-from .song.song_detail_crawler import crawlSongDetail
+from ..config import GenieConfig
+from .utils import getLogger
+from .song import crawlSongDetail
 
 SONG_DETAIL_URL = "https://www.genie.co.kr/detail/songInfo?xgnm="
 config = GenieConfig()
+
 
 def crawlSong(song_ids: list[str]) -> pd.DataFrame:
     # log file setting
@@ -31,5 +32,5 @@ def crawlSong(song_ids: list[str]) -> pd.DataFrame:
         log.warn("Failed crawling song ids : %s", str(failed_crawling_song_ids))
 
     song_details_df = pd.DataFrame(song_details)
-    
+
     return song_details_df
