+++ src/migration/__init__.py
@@ -1,0 +1,5 @@
+from .crawler import CrawlerMigrate
+from .playlist import PlaylistCsvMigrate
+from .song import SongCsvMigrate
+
+__all__ = ["CrawlerMigrate", "PlaylistCsvMigrate", "SongCsvMigrate"]
